Add a readme
This is an empty repository. This adds a readme that spells where the project is
going and why it exists. It also points to some related projects and
documentation that will be helpful once more of it is up and running.
ActiveSupport carries a number of useful tools for mangling strings into formats
often used by programming languages. ERB will be the templating langauge used
inside documents and for transforming file names. That filename part might end
up being regrettable but for now it'll do.