Add Gemfile
Adds a minimal gemfile, defining only a Rake version necessary to get
the application running. Maybe that should be a gemspec dependency? All
other dependencies are defined in the gemspec file. It uses rubygems.org
as a source for gems although repeatable builds dictate that eventually
this should move to vendored gems and a privately maintained source for
gems. For an application used by others there's an argument to be made
that a well respected public source for gems is better.