Modify gemspec to use Library Version
By using the library's version constant in the gemspec declaration there
is no chance that a new gem version will be pushed that doesn't match
the one declared in the library.

A rake task to "bump the version" and push a new gem should be written
to further reduce the chance of a release without version bump.