Initialize Tapioca and Configure Sorbet
Tapioca is replacement tooling for Sorbet's signature discovery. These
files were added by running `bundle exec tapioca init` to stub out an
application in /bin and setup some configuration files. The sorbet
config will be modified in a separate commit to ignore bin/ because it
isn't really part of this paplication and test/ to avoid noise.

Configure Sorbet

Have sorbet explicitly ignore the contents of bin/ and test/ because
they aren't really part of the application

Ignore sorbet/ folder

The sorbet folder holds signatures generated by tapioca to provide
type-checking for the application. This is tooling-generated data that
should be unique to each machine as gemfile.lock isn't stored, gems
aren't vendored and versions aren't locked in gemspec.

The "getting started with development" documentation should be modified
to recommend running `bundle exec tapioca gem --workers=16` to generate
these definitions in parallel, then run `bundle exec tapioca todo` to
create the "todo list" for gems which don't have signatures.

Perhaps a task in Rakefile and in bin/setup should be added too. Revisit
the choice to ignore this directory in the future.