Add License to yard documentation
The Rakefile was already configured to include license.txt even though
the file didn't exist. This adds it to .yardopts so even if
documentation is generated manually it'll be included. Run `rake doc` to
regenerate documentation and ensure that the "license" is included in
the files section of the generated docs.