Add App class
The application class will be the main entry point for the program. You
will instantiate it, configure some settings like paths for templates,
dictionaries of variables, and other run-time behaviour and then call
"run" to make it all happen.

For now this just plumbs in the "run" method to say hello and wires up
some tests for ensuring future behaviour. It also adds it to the
library. The next step is to write an "exe" that instantiates App and
calls run so that `bundle exec exe/project_templates` actually does
something. That's the last essential bit of plumbing before the program
can do something useful.