Add main executable
Adds the main executable of the application. Once more of the
application is built this will be extended to pass along command line
arguments. For now it instantiates and runs the application which will
print "hello".

You can run the program with `bundle exec exe/project_templates`. The
readme should be updated to account for the fact this gem does something
now. Some github actions to ensure tests are run automatically are a
good idea as is configuring build & release.

It may be time to revisit the bin/setup script too.