Create config class
The configuration class is the container for all of the settings that are
necessary to run the application. The expected use is to create a new
configuration using one of the class methods though they don't exist right now.
In the interim you can specify all of the values during initialization. Maybe
that'll be with a block, maybe that'll be be with a million params on
`initialize`. It's not clear yet.

Configs can be passed to the application which will delegate all of the lookups
of config values back to the configuration object.