App learns about config
The App class is now able to be passed a config. The config should be
setup and passed in ready-to-use when initializing the application. A
new CLI class will need to be created; that will read in command line
arguments, environment variables, etc, and create a config. It will then
instantiate an application and call run. This approach will allow
different user interfaces to use the same configuration and app class.