exe/project_templates learns about config
The main executable now creates a config that is passed into the
application it uses to "do work". The end game here is to have it
instantiate a CLIApp class which wraps app, builds a config, and exposes
'run' of its own.

This change is mostly done to facilitate testing/debugging.