Improve testing with `focus` and `pry`
This adds some improvements to the test/debut cycle.

  1. It adds minitest-focus which allows you to write "focus" above the
     definition of a test and then only that test will run. For more
     details read https://github.com/seattlerb/minitest-focus

  2. Adds pry to the test helper. This allows you to add `binding.pry`
     to a test and be dumped into a debugger when you reach that point.

While we're here we also use the MiniTest::Mock library that was added
in 55c993b2 to mock a config rather than passing an explicit nil.
Strictly speaking that isn't required for this but it does kinda glue
together all these improvements that are being made.