Some style improvements in tests and in rubocop rules
This change ensures that when using multi-line arrays or hashes that the
last line has a column. Effective the following definitions are okay:

```
  dict = { foo: 1, bar: 2 }
  dict = {
    foo: 1,
    bar: 2,
  }
```
and likewise for arrays:
```
  arr = [1, 2]
  arr = [
    1,
    2,
  ]
```
This change requires an update to the gemspec to be valid.

Adding some brackets around arguments to assertions. Just to be a little
more consistent with the rest of the code. Some minor changes to make
test naming more consistent by dropping "it".