Build path attributes for config
This adds the most important paths as attributes for config.

* `template_path` is the location of project templates
  By default imagine something like ~/.local/share/project_templates

* `project_path` is the location of the specific project template to be
  used as a source. By default imagine something like
  ~/.local/share/project/templates/ruby_application

  in this example `project_name` would be `ruby_application` and it was
  appended to `template_path` to compute this path.

* `target_path` is the destination where the template will be generated.
  by default this would be a path in the current directory plus
  `target_name`. Something like ~/src/github.com/me/cool_app if
  `target_name` was "cool_app" and the current working directory was
  ~/src/github.com/me/

It also defines a few DEFAULT_ constants. These might have to go and we'll just
make everything "mandatory". I'll figure that out once all the
attributes are in place.

Finally, there's some minor test clean up to use the new helpers and to
simplify initialization: one for "it uses defaults" and one for "it uses
values passed in".