Rubocop style updates
Allow test methods to break some rules. Tests get to do multiple
assignments because it makes reading them easier, plus assertions
sometimes need to happen on multiple bits of data in a single test. The
ABC cop is disabled for test/**/*

This is also the reason the line length for methods is disabled:
sometimes a test really should do a lot of stuff.

The rubocop file was also reordered to be alphabetical.