Config learns working_path, project, and target name
Continuing on the the building of the Config object.

* `working_path` is defined as an attribute. This is the location
    where the output of the template will be written. It will be
    prefixed to `target_name` to produce `target_path`.

  * `target_name` was added: it's the name of the new directory to
    create inside of `working_path`.

  * `project_name` was added: it's the name of the project to use as the
    source. It's appended to `template_path` to produce `project_path`

Default constants for the above are defined. Again, just placeholders
that will likely change.

Alias for `dry_run?` moved closer to the `dry_run` accessor declaration
just because it makes sense.

the `initialize()` method is growing pretty heavy on arguments. I'm
disabling the warning cop for now. Longer term it seems like a `paths`
variable to hold `template`, `working`, `target`, and `project` might be
a good idea.

A similar need will arise for the "dictionaries" of variables that will
be provided for command line, project, and global defaults.

`paths`, `variables`, `target_name`, `dry_run` are probably all we
really want for `the `initialize` method.