Stub in a Dictionary class
The `Dictionary` will hold variables. These may be from a YAML file,
JSON, or just "explicitly set" with some kind of a `config {}` block.
I'm not totally sure what this looks like.

There will be instances of dictionaries for "global" variables from
~/.config/project_templates/variables.yaml, "project" variables from
~/.local/share/project_templates/project/foo_project/variables.yaml, and
one to wrap arguments provided as command line arguments.

All of these dictionaries will be "wrapped" in a variables object which
will be responsible for enforcing the override priorities: command line
> project > global.

Maybe this will will just be a glorified openstuct.  The goal is to have
an interface that can be navigated with "dots" in templates.
`foo.bar.baz` rather than `config[:foo][:bar][:baz]`.