Add class_under_test helper module
There was already a bit of a naive version of this method defined in the
`TestConfig` class by explicitly declaring an instance variable and
adding an attr_reader. Instead of doing that in every class we've got a
simple helper that can be 'included' into a test file.

The implementation is trivial: map `TestFoo` to `ProjectTemplates::Foo`.
It doesn't try to do any clever path lookup or walking of the object
hierarchy: just straight string manipulation. It won't be foolproof, but
it's simple enough to understand and makes the tests a lot easier to
read and write.