Implement Dicionary parsing
A Dicionary class now exists. It can be instantiated with
  * A ruby hash
  * An OpenStruct
  * A string of JSON
  * A string of YAML
Failure to parse the input will raise argument error.

A Dicionary instance delegates all method calls to the struct that it
was instantiated with: each of the keys becomes a method call that
returns its corresponding value. This Key-to-method-call trick is
performed recursively.

```
dict = Dictionary.load({a: { b: 1}, c: 2 })
dict.a.b # => 1
dict.c # => 2
dict.d # raises NoMethodError
```

This is intended to wrap input loaded from files and also arguments
passed as command line attributes.