Handle "invalid" ruby hash passed to dict
In order to create a Dictionary from a ruby hash all keys must be
symbolizable. This is because a dictionary uses keys to define methods
(or more precisely to delegate methods to an openstruct). If Keys can't
be symbolized method naming all falls apart.

This gets handled during the `load` phase but the error is confusing:
`NoMethodError` for `to_sym` on a hash key that can't be symbolized.

This swallows that error and bubbles it up as an ArgumentError which is
more consistent with the rest of the interface.