Wire Dictionaries into Config
The config object now knows about global, project, and run variables. It
can expose these to the application which can use them to populate
templates.

The default values that are fed into dictionaries aren't very useful.
The next step is to modify the executable to "configure" the Config by
loading dictionaries and also setting things like the paths and names
appropriate.