Notes for config setup at run time
The entire config object needs to be built out. In order to make that
happen environment variables, files, and command line arguments need to
be parsed. Some settings need to be computed based on earlier
declarations. Some settings need to override other settings.

The exact structure of this is not yet fully determined. This serves as
some notes for how it might go but the details will be worked out
through experimentation.