Continue WIP config loading
Pushing forward on the concept of a "config loader". The idea is that
you create a loader which returns a dictionary. Loaders are configured
with a series of sources which it will try in order. Those sources are
strings: either a path to a file, or a raw sting that can be consumed.

I'm not sure exactly what this API should look like. I do think it
should be either you get back a dictionary every time. I also think the
loader should be queryable for success and the the validity of all
sources should be something you can look at too. Did the source fail
because the path didn't exist or because the content at that path was
useless?

I know the class method for instantiating a lower here sucks. I probably
want an abstract class that defines the interface, or maybe a mix in for
each of the major types of loader. Wanting to support environment
variables makes this harder - env-as-path and env-as-string. also
relative vs absolute path...

Being able to use other dictionaries as a source would be nice too.
Global variables should be able to overide ones in projects, but also
let you define other files to read...maybe that's a later problem after
the basic work is done