Work in progress - building template sources
I'm going to owe a bunch of tests, a better test helper, and some
refactoring of file checking dance. I source also delegate rather than
call @path for things. There might be a way to re-write ENV sources in
terms of the non-env counterparts.

Committing this as a WIP branch just to not accidentally lose progress.

The plan here is to simplify dictionary: it just comes from a JSON/YAML
string. The loaders are responsible for getting that from somewhere.

Maybe it was a bit overbuilt.