Work in Progress - cleaned up loaders
Continued work on these to be a little less gross.
Dictionary is trimeed, config updated to use string
but it'll probably get changed to use a loaders instead (the empty one
for default).

Still needs work but we're making good progress. IOU some tests which
means I need to figure out the best way to mock file system and env.