Work in Progress
Continue work on the loaders. There is now a simple environment mock
method in the test helper. Growing a lot of modules here so should
probably look at consolidating that stuff. It's annoying to have to
include common features like that.

Tests are coming along. Need a way to mock the file system.
still having some weirdness with defining env files too.

Once file loading is set we can move on to config refactor...