Work in Progress - Dictionary loading tests
At this point all of the loaders are loading and testing. There are
reasonable helpers for making sure these tests are easy to write and
maintain.

The last round of updates will be to make loaders swallow up the
exception and return a nil dictionary. This should be possible to cover
in the `load` method. Will probably need some new interface stuff like a
`#success?` or `#faiure_message` but that can wait until they're
actually needed.

With this settled we should be at the point to modify the exe file to
load the initial config file.