Stub out a Dictionary Loader.
The dictionary loader takes in a bunch of possible dictionary sources
and either gives you back a dictionary from one of those sources or
it "blows up" and you get nil back.

The exact use of these things is yet to be determined.

Should this be abstract and you make specific dictionary loaders
:

```
class MainConfigLoader
  extend DictionaryLoader

  SOURCES = [
    Sources::EnvString.new("PROJECT_TEMPLATES_CONFIG")
    Sources::EnvFile.new("PROJECT_TEMPLATES_CONFIG_FILE"),
    Sources::File.new("~/.config/project_templates/config.yaml"),
    Sources::Empty.new
  ]

  # some nice version of
  def initialize(sources = SOURCES)
    super(sources)
  end
end
```

It's convenient to be able to just 'make a loader' so maybe this is
an abstract version or a base class and then the main config, project
and args versions specialize?

Maybe a factory:

```
class ConfigFactory
  def self.main_config
    DictionaryLoader.new(
      Sources::EnvString.new("PROJECT_TEMPLATES_CONFIG")
      Sources::EnvFile.new("PROJECT_TEMPLATES_CONFIG_FILE"),
      Sources::File.new("~/.config/project_templates/config.yaml"),
      Sources::Empty.new
    )
  end
  ...
end
```
That feels a bit "Java".

We could get all meta and use a config loader to read a file
that we use to define the sources for config loaders: but that's
probably a bridge too far.

I'm going to hold off for now and see how it feels later.

Also: Add a bunch of comments to sources to describe what they do
so there are some nicer docs