Namespace Aware class_under_test helper
This still isn't perfect, but it beats pulling active support for just
this little thing. It figures out what class a test maps back to. It
does that by assuming the name of a test maps to a class:

1. `/test/foo/bar/test_baz.rb` defines `TestBaz`
2. `TestBaz` is the tests for the `Baz` class.
3. `/lib/foo/bar/baz.rb` defines `Baz`

class_under_test looks at the class for the current test, backs up the
directory to the root of the project, does some directory math to figure
out the path to the current test file. Then it does some basic string
munging to try and make that path into what the probable namespaces are.
Finally, it mangles the test class in a way that it should match the
tested class and glues that to the end of the computed namespace.

If all this goes wrong you can always define @class_under_test
= Foo::Bar:baz in the test's setup block and that will be used instead