Work in Progress - Build out Config Paths
The `Config` object has a lot of accessors and the initializer is pretty
complicated. Let's group all of the paths together in a sub-object so
that `Config.paths` can be meaningful. This object will compute (or
allow you to explicitly set) paths that will be used for writing out a
new project from a template.

It also does basic validation to ensure everything can be used.