Work in Progress - Rework Dictionaries
So now dictionaries are just normal classes. The input
is stored as useful ruby hash in addition to the openstruct

Now you can `load` it from YAML or `JSON` or `new` it form a json string
and either you get an argument error because it doesn't produce a
dictionary or you get a thing you can call methods on.

A bunch of tests will need updating now